feat: 소설 선택 기능 구현
