+++ app/src/main/res/layout/item_search_novel.xml
@@ -8,11 +8,16 @@
         <variable
             name="novel"
             type="com.teamwss.websoso.ui.normalExplore.model.NormalExploreModel.NovelModel" />
+
+        <variable
+            name="onClick"
+            type="kotlin.jvm.functions.Function1" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:onClick="@{() -> onClick.invoke(novel.id)}"
         android:paddingTop="6dp">
 
         <ImageView
@@ -67,8 +72,8 @@
             android:layout_width="24dp"
             android:layout_height="24dp"
             android:layout_marginEnd="10dp"
-            android:padding="10dp"
             android:background="@drawable/img_withdraw_second_check_button_selector"
+            android:padding="10dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
